requirements.txt
fastapi
uvicorn
jinja2
pillow
numpy
scikit-learn
tensorflow
joblib
python-multipart
supabase
python-dotenv
python-jose





